Move accumulations to a different package
