+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/aggregator/Aggregator.java
@@ -5,7 +5,7 @@
 
 package io.opentelemetry.sdk.metrics.aggregator;
 
-import io.opentelemetry.sdk.metrics.aggregation.Accumulation;
+import io.opentelemetry.sdk.metrics.accumulation.Accumulation;
 import javax.annotation.concurrent.Immutable;
 
 /**
