+++ sdk/metrics/src/test/java/io/opentelemetry/sdk/metrics/accumulation/LongAccumulationTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.metrics.aggregation;
+package io.opentelemetry.sdk.metrics.accumulation;
 
 import static org.assertj.core.api.Assertions.assertThat;
