+++ sdk/metrics/src/test/java/io/opentelemetry/sdk/metrics/aggregator/DoubleLastValueAggregatorTest.java
@@ -7,7 +7,7 @@ package io.opentelemetry.sdk.metrics.aggregator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.opentelemetry.sdk.metrics.aggregation.DoubleAccumulation;
+import io.opentelemetry.sdk.metrics.accumulation.DoubleAccumulation;
 import org.junit.jupiter.api.Test;
 
 /** Unit tests for {@link AggregatorHandle}. */
